+++ pyvela/pyvela/pyvela_plot.py
@@ -44,11 +44,14 @@ def get_param_plot_mask(
     ]
 
 
-def read_true_values(result_dir: str) -> np.ndarray:
+def read_true_values(result_dir):
     with open(f"{result_dir}/summary.json", "r") as summary_file:
         summary = json.load(summary_file)
 
-    if "truth_par_file" not in summary["input"]:
+    if (
+        "truth_par_file" not in summary["input"]
+        or summary["input"]["truth_par_file"] is None
+    ):
         return None
 
     true_values_raw = np.genfromtxt(f"{result_dir}/param_true_values.txt")
@@ -61,6 +64,7 @@ def plot(
     result_dir: str,
     ignore_params: Iterable[str] = [],
     include_params: Iterable[str] = None,
+    plot_priors: bool = False,
     out: str = None,
     labelpad: float = 0.2,
 ):
@@ -113,6 +117,16 @@ def plot(
         ax.tick_params(axis="both", labelsize=8)
         ax.yaxis.get_offset_text().set_fontsize(8)
         ax.xaxis.get_offset_text().set_fontsize(8)
+
+    if plot_priors:
+        # Plot the pre-evaluated priors
+        prior_evals = np.load(f"{result_dir}/prior_evals.npy")
+        nplots = len(param_plot_mask)
+        for jj, ii in enumerate(param_plot_mask):
+            plt.subplot(nplots, nplots, jj * (nplots + 1) + 1)
+            xs = prior_evals[:, 2 * ii]
+            ys = prior_evals[:, 2 * ii + 1]
+            plt.plot(xs, ys)
 
     ax = plt.subplot(5, 2, 2)
     ax.errorbar(