+++ src/prior/prior_scaling.jl
@@ -1,5 +1,3 @@
-export scale_prior_args
-
 prior_scaling(::Type{Arcsine}, args) = map(oneunit, args)
 
 prior_scaling(::Type{Beta}, args) = map(zero, args)
@@ -77,3 +75,6 @@ prior_scaling(::Type{Weibull}, args) = ((), (0,), (0, 1))[length(args)+1]
 
 scale_prior_args(distr_type, args, unit_conversion_factor) =
     args .* (unit_conversion_factor .^ prior_scaling(distr_type, args))
+
+unscale_prior_args(distr_type, args, unit_conversion_factor) =
+    args .* ((1 / unit_conversion_factor) .^ prior_scaling(distr_type, args))