+++ CHANGELOG.md
@@ -12,14 +12,24 @@
 - Can save the plot output from `pyvela-plot`
 - Can specify the label padding in `pyvela-plot`
 - Can specify only the parameters to include in `pyvela-plot`
+- `PriorSourceType` enum, `source_type` attribute in `Prior` types
+- `SPNTA.full_prior_dict()` and `SPNTA.save_results()` methods
+- `unscale_prior_args()` function
+- Save the full prior information along with the results
+- Convert `RNAMP` and `RNIDX` to `TNREDAMP` and `TNREDGAM`
+- Plot priors in `pyvela-plot`
 ## Changed
 - Residual plot in `pyvela-plot` script
 - Made GP noise marginalization the default in `SPNTA`
 - Made plotting modular
+- Residual plot in `pyvela-plot` script
+- Made GP noise marginalization the default in `SPNTA`
+- Move `info_dict()`, `save_new_parfile()`, `save_resids()` from `pyvela_script` to `SPNTA`
 ## Fixed
 - Julia example script `run_example.jl`
 - In `_gls_lnlike_serial`, return  -inf if `Σinv` is not positive definite.
 ## Removed
+- Saving maximum-posterior par file in `pyvela` script
 
 # 0.0.9
 ## Added